requirements.txt
django-model-utils==4.0.0
Django==3.1
django-bootstrap4==2.2.0
django-crispy-forms==1.9.2
pytest-django==3.10.0
djangorestframework==3.11.0
